plugins/aws-agentcore-langgraph/examples/memory-demo/pyproject.toml
[project]
name = "memory-demo"
version = "0.1.0"
description = "AWS Bedrock AgentCore Memory demonstration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bedrock-agentcore>=1.0.7",
    "boto3>=1.35.0",
    "langchain>=1.1.0",
    "langchain-aws>=1.1.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
]
